Fix broken quick form test after updating code to use field getters and setters
Currently the tests are relying on a $entity->quick dynamic property
to be created and saved on the entity object which is then passed back
to the tests via form state storage. Instead, we need to make sure we are using
the real entity.quick field we provide. This required adding farm_entity and
entity module dependencies and implementing plugins for the entity bundles.

Fixes Drupal\Tests\farm_quick\Kernel\QuickFormTest::testQuickFormSubmission
InvalidArgumentException: Field quick is unknown.