Do not extend Uri constraint from Symfony's Url constraint.
We provide our own UriValidator that does not extend from
Symfony's UrlValidator, so there's no reason this needs to
be related.

This fixes the following deprecation notice in tests:

Since symfony/validator 7.1: Not passing a value for the "requireTld"
option to the Url constraint is deprecated. Its default value will
change to "true".